requirements.txt
requests
pandas
sqlalchemy
python-dotenv
beautifulsoup4
loguru
playwright
